+++ kernel/src/shell/cmds.rs
@@ -39,3 +39,4 @@ pub mod write;
 pub mod pci;
 pub mod grep;
 pub mod play;
+pub mod hda;